% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kendallknight-package.R
\docType{data}
\name{arcade}
\alias{arcade}
\title{Number of doctorates versus arcade revenue in the United States}
\format{
A data frame with 10 rows and 3 variables:
\describe{
\item{year}{Year of the observation.}
\item{doctorates}{Number of doctorates awarded in computer science.}
\item{revenue}{Total revenue generated by arcades (in billions of dollars).}
}
}
\source{
Spurious Correlations (Vigen 2015)
}
\usage{
arcade
}
\description{
A dataset containing the yearly number of doctorates awarded in computer
science and the total revenue generated by arcades in the United States
for the period 2000-2009.
}
\examples{
arcade
}
\keyword{datasets}
