% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kendallknight-package.R
\docType{data}
\name{cigarettes}
\alias{cigarettes}
\title{Life expectancy and cigarettes per day}
\format{
A data frame with 15 rows and 2 variables:
\describe{
\item{life_expectancy}{Life expectancy in years.}
\item{cigarettes_per_day}{Cigarettes smoked per day.}
}
}
\source{
Real Statistics Using Excel (\url{https://real-statistics.com/correlation/kendalls-tau-correlation/kendalls-correlation-testing-with-ties/}).
}
\usage{
cigarettes
}
\description{
A dataset containing life expectancy and cigarettes per day.
}
\examples{
cigarettes
}
\keyword{datasets}
