% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kendallknight-package.R
\docType{package}
\name{kendallknight-package}
\alias{kendallknight}
\alias{kendallknight-package}
\title{kendallknight: Efficient Implementation of Kendall's Correlation Coefficient Computation}
\description{
The computational complexity of the implemented algorithm is \eqn{O(n \log(n))}{O(n log(n))}, which is faster than the base R implementation in \code{\link[stats]{cor}} with \code{method = "kendall"} that has a computational complexity of \eqn{O(n^2)}{O(n^2)}. For small vectors (i.e., less than 100 observations), the time difference is negligible. However, for larger vectors, the difference can be substantial. The references are Knight (1966) \doi{10.2307/2282833}, Abrevaya (1999) \doi{10.1016/S0165-1765(98)00255-9}, Christensen (2005) \doi{10.1007/BF02736122} and Emara (2024) \url{https://learningcpp.org/}. This implementation is described in Vargas Sepulveda (2024) \doi{10.48550/arXiv.2408.09618}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://pacha.dev/capybara/}
  \item \url{https://github.com/pachadotdev/capybara}
  \item Report bugs at \url{https://github.com/pachadotdev/capybara/issues}
}

}
\author{
\strong{Maintainer}: Mauricio Vargas Sepulveda \email{m.sepulveda@mail.utoronto.ca} (\href{https://orcid.org/0000-0003-1017-7574}{ORCID})

Other contributors:
\itemize{
  \item Loader Catherine (original stirlerr implementations in C (2000)) [contributor]
  \item Ross Ihaka (original chebyshev_eval, gammafn and lgammacor implementations in C (1998)) [contributor]
}

}
